Fix compilation of the codesign program
I'm not familiar with the `System.CommandLine` API but it seems something has changed recently, see [Announcing System.CommandLine 2.0 Beta 2 and the road to GA][1]. I think this happened because the `System.CommandLine` package is referenced with a floating version (2.0.0-*) and a breaking change was introduced.

Anyway, `CommandHandler.Create` has been moved into a new `System.CommandLine.NamingConventionBinder` compatibility NuGet package and the new way of setting up a command handler is with a new `SetHandler` extension method.

[1]: https://github.com/dotnet/command-line-api/issues/1537