Fix rewriting of the __LINKEDIT section.
There was an incorrect assumption that __LINKEDIT section can only
change when contents of MachLinkEditData change. However, that's not
the only case when that can happen. It also changes when load commands
are added or removed.

Added basic test for code signing an a.out file to ensure that it does
not regress.